Now efficiently computing a bootstrap sample of full AUC from controls and cases with bootstrapAucStratified

+++ src/RcppExports.cpp
@@ -6,7 +6,7 @@
 using namespace Rcpp;
 
 // bootstrapAucStratified
-std::vector<double> bootstrapAucStratified(const std::vector<double>& controls, const std::vector<double>& cases, const int bootN);
+std::vector<double> bootstrapAucStratified(const std::vector<double>& controls, const std::vector<double>& cases, const size_t bootN);
 RcppExport SEXP pROC_bootstrapAucStratified(SEXP controlsSEXP, SEXP casesSEXP, SEXP bootNSEXP) {
 BEGIN_RCPP
     SEXP __sexp_result;
@@ -14,7 +14,7 @@ BEGIN_RCPP
         Rcpp::RNGScope __rngScope;
         Rcpp::traits::input_parameter< const std::vector<double>& >::type controls(controlsSEXP );
         Rcpp::traits::input_parameter< const std::vector<double>& >::type cases(casesSEXP );
-        Rcpp::traits::input_parameter< const int >::type bootN(bootNSEXP );
+        Rcpp::traits::input_parameter< const size_t >::type bootN(bootNSEXP );
         std::vector<double> __result = bootstrapAucStratified(controls, cases, bootN);
         PROTECT(__sexp_result = Rcpp::wrap(__result));
     }
